Fix XP overflow, harden the leveling curve, add 4 endgame zones (100-500)
XP curve overhaul:
- Widen currentXp/xpToNextLevel (Stats.java) and XpCurve.Entry's
  xpToNextLevel/totalXp from int to long, plus every call site
  (PlayerProgression, GmCommandProcessor, DeathPenaltyService,
  PlayerStateDto, T4CContentStudio). The old int totalXp silently
  overflowed negative starting at level 541 in the committed curve data.
- Replace the old curve generator (XpCurveExtender, int-accumulating,
  hence the overflow) with XpCurveHardener: levels 1-99 unchanged,
  100-1000 use an exponent that itself grows with level, so the curve
  structurally outgrows any fixed-degree monster-XP scaling (verified
  against the whole monster roster - worst remaining single-kill skip
  anywhere is ~3.7 levels, from three likely-intentional one-time
  "world boss" kills, down from unbounded/overflowed before).

Also fixes Mirak Nira's "kill 100 goblins" trust quest, which was
silently unwinnable - nothing ever incremented the counter it reads.
Any GOBLIN-clan kill, anywhere, now counts.

Four new level 100-500 zones, calibrated against the new curve:
- Windhowl Marches (100-150): Centaur Warrior/King, previously
  barely-authored placeholder stubs, given real stats and spawns.
- The Hollow March (150-200): new Barrow Wight / The Hollow King.
- Lesser Drake's Aerie (200-260): new Kraanian Wyrmling + activates
  the dormant, fully-statted-but-never-placed Lesser Drake (xpOnDeath
  was 0, zero spawns).
- Greater Drake's Bastion (400-500): new Bastion Warden + activates
  Greater Drake the same way.

Each zone: 2 items, 1 quest + quest-giver NPC. Full writeup, including
the linear-extrapolation miscalibration caught and fixed mid-pass for
the Bastion Warden, in docs/content-ideas/.

+++ src/main/java/com/perso/T4C/helper/XpCurve.java
@@ -36,12 +36,12 @@ public final class XpCurve {
     return Collections.unmodifiableList(entries);
   }
 
-  public int getXpToNextLevel(int level) {
+  public long getXpToNextLevel(int level) {
     Entry entry = byLevel.get(level);
     return entry != null ? entry.xpToNextLevel : 0;
   }
 
-  public int getTotalXp(int level) {
+  public long getTotalXp(int level) {
     Entry entry = byLevel.get(level);
     return entry != null ? entry.totalXp : 0;
   }
@@ -55,7 +55,7 @@ public final class XpCurve {
       level = 1;
       player.setLevel(level);
     }
-    int xpToNext = getXpToNextLevel(level);
+    long xpToNext = getXpToNextLevel(level);
     if (xpToNext > 0) {
       player.setXpToNextLevel(xpToNext);
     }
@@ -70,10 +70,10 @@ public final class XpCurve {
   @Getter
   public static class Entry {
     private final int level;
-    private final int xpToNextLevel;
-    private final int totalXp;
+    private final long xpToNextLevel;
+    private final long totalXp;
 
-    public Entry(int level, int xpToNextLevel, int totalXp) {
+    public Entry(int level, long xpToNextLevel, long totalXp) {
       this.level = level;
       this.xpToNextLevel = xpToNextLevel;
       this.totalXp = totalXp;
