+++ src/main/java/com/perso/T4C/player/PlayerProgression.java
@@ -37,7 +37,7 @@ public final class PlayerProgression {
       player.setSkillPoints(player.getSkillPoints() + 15);
       player.showLevelUpMessage(player.getLevel(), gains[0], gains[1]);
       player.notifyLevelUp(player.getLevel());
-      int next = xpCurve.getXpToNextLevel(player.getLevel());
+      long next = xpCurve.getXpToNextLevel(player.getLevel());
       if (next > 0) {
         player.setXpToNextLevel(next);
       } else {
